edit update delete

+++ src/jobportal/UserActionServlet.java
@@ -12,7 +12,7 @@ import javax.servlet.http.HttpServletResponse;
 /**
  * Servlet implementation class UserActionServlet
  */
-@WebServlet("/deleteuser")
+@WebServlet("/useraction")
 public class UserActionServlet extends HttpServlet {
 	private static final long serialVersionUID = 1L;
        
@@ -28,9 +28,22 @@ public class UserActionServlet extends HttpServlet {
 	 * @see HttpServlet#doGet(HttpServletRequest request, HttpServletResponse response)
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-		String userId =request.getParameter("userId");
-		if(userId != null){
-			DBUtils.deleteUser(Integer.parseInt(userId));
+		String action = request.getParameter("action");
+		if(action.equals("delete")){
+			request.setAttribute("action", "delete");
+			String userId =request.getParameter("userId");
+			if(userId != null){
+				DBUtils.deleteUser(Integer.parseInt(userId));
+			}
+		} else if( action.equals("edit")){
+			request.setAttribute("action", "edit");
+			User user = new User();
+			String userId =request.getParameter("userId");
+			if(userId != null){
+				user = DBUtils.getUser(Integer.parseInt(userId));
+			}
+			
+			request.setAttribute("user", user);
 		}
 		List<User> userList = DBUtils.getUsers();
 		request.setAttribute("users", userList);
